+++ tigeropen/common/consts/service_types.py
@@ -31,6 +31,7 @@ SEGMENT_FUND_AVAILABLE = "segment_fund_available"
 TRANSFER_SEGMENT_FUND = "transfer_segment_fund"
 CANCEL_SEGMENT_FUND = "cancel_segment_fund"
 PLACE_FOREX_ORDER = "place_forex_order"
+ESTIMATE_TRADABLE_QUANTITY = "estimate_tradable_quantity"
 
 USER_LICENSE = "user_license"
 
@@ -60,6 +61,7 @@ QUOTE_STOCK_TRADE = "quote_stock_trade"
 QUOTE_DEPTH = "quote_depth"  # level2 深度行情
 GRAB_QUOTE_PERMISSION = "grab_quote_permission"  # 抢占行情
 MARKET_SCANNER = "market_scanner"  # 选股器
+MARKET_SCANNER_TAGS = "market_scanner_tags"
 GET_QUOTE_PERMISSION = "get_quote_permission"
 TRADING_CALENDAR = "trading_calendar"
 STOCK_BROKER = "stock_broker"  # 港股股票实时经纪队列
@@ -67,6 +69,7 @@ CAPITAL_DISTRIBUTION = "capital_distribution"  # 股票当日资金分布
 CAPITAL_FLOW = "capital_flow"  # 股票资金流向
 WARRANT_FILTER = "warrant_filter"
 WARRANT_REAL_TIME_QUOTE = "warrant_real_time_quote"
+KLINE_QUOTA = "kline_quota"  # 历史k线额度
 
 # 期权行情
 OPTION_EXPIRATION = "option_expiration"