+++ tigeropen/trade/domain/contract.py
@@ -17,7 +17,7 @@ class Contract:
                  market=None, min_tick=None, trading_class=None, status=None, continuous=None, trade=None,
                  marginable=None, close_only=None,
                  last_trading_date=None, first_notice_date=None, last_bidding_close_time=None, tick_sizes=None,
-                 is_etf=None, etf_leverage=None):
+                 is_etf=None, etf_leverage=None, **kwargs):
         self.contract_id = contract_id
         self.symbol = symbol
         self.currency = get_enum_value(currency)
@@ -83,6 +83,12 @@ class Contract:
         # ETF杠杆倍数
         self.etf_leverage = etf_leverage
 
+        self.discounted_day_initial_margin = kwargs.get('discounted_day_initial_margin')
+        self.discounted_day_maintenance_margin = kwargs.get('discounted_day_maintenance_margin')
+        self.discounted_time_zone_code = kwargs.get('discounted_time_zone_code')
+        self.discounted_start_at = kwargs.get('discounted_start_at')
+        self.discounted_end_at = kwargs.get('discounted_end_at')
+
     @property
     def right(self):
         return self.put_call
@@ -103,3 +109,48 @@ class Contract:
         :return:
         """
         return self.sec_type == 'STK' and (self.currency == 'CNH' or self.currency == 'CNY')
+
+
+class ContractLeg:
+    def __init__(self, symbol=None, sec_type=None, expiry=None, strike=None, put_call=None, action=None,
+                 ratio=1):
+        self.symbol = symbol
+        self.sec_type = sec_type
+        self.expiry = expiry
+        self.strike = strike
+        self.put_call = put_call
+        self.action = action
+        self.ratio = ratio
+
+    def to_openapi_dict(self):
+        d = dict(**self.__dict__)
+        d['right'] = d.pop('put_call', None)
+        return d
+
+    def __repr__(self):
+        return str(self.__dict__)
+
+class OrderContractLeg(ContractLeg):
+    def __init__(self, symbol=None, sec_type=None, expiry=None, strike=None, put_call=None, action=None,
+                 ratio=1, market=None, currency=None, multiplier=None, total_quantity=None, filled_quantity=None,
+                 avg_filled_price=None, created_at=None, updated_at=None, **kwargs):
+        if 'right' in kwargs and not put_call:
+            put_call = kwargs.get('right')
+        super().__init__(symbol=symbol, sec_type=sec_type, expiry=expiry, strike=strike, put_call=put_call, action=action,
+                 ratio=ratio)
+        self.market = market
+        self.currency = currency
+        self.multiplier = multiplier
+        self.total_quantity = total_quantity
+        self.filled_quantity = filled_quantity
+        self.avg_filled_price = avg_filled_price
+        self.created_at = created_at
+        self.updated_at = updated_at
+
+    def to_openapi_dict(self):
+        d = super().to_openapi_dict()
+        d.update(self.__dict__)
+        return d
+
+    def __repr__(self):
+        return str(self.__dict__)